Add decent CSS elements.Improve display.
-The calculator now has two displays so that the user can see
both numbers on the screen. To make it possible it was necessary
to add a new html elemnt ( div class='second-display'), modify all
functions ( calculator(), del(), clear()) alongside with first display
behavior. Second display keeps evidence of the firstNum and operator
while on the other user sees secondNum or the result of one operation
(***The result is displayed when the second operator is clicked)
**operate() is now deleted because was not useful at all.

TO DO: add keyboard support.