Stop text from overflowing. Fix addDecimals().
- In addDecimals(), 2nd if() statement had a variable named
decimalPressed. It was supposed to be named firstNumDecimal. This
error cased a bug where the user could add more than one 'dots'
to a single number.
- Add in reset() function the initial state(=false) of the variables
responsible for adding decimals to numbers
- Add an aditional if() statemen in del() function to check if the
deleted character is a dot: if it's true it sets the responsible
variables to their initial state.
- Stop the text to overflow the container by adding overflow: hidden
CSS property.